+++ app_server/controllers/travel.js
@@ -2,6 +2,7 @@
 module.exports.travel = function(req, res) {
     res.render('travel', { 
         title: 'Travel - Travlr Getaways',
+        page: 'travel',
         layout: 'layouts/layout'
     });
 };